feat(web): add verification to prevent access private page when users is not authenticated
